Rename CavesChunkGeneratorMixin.java to ChunkGeneratorSettingsMixin.java as it isn't just the caves chunk generator anymore.
